services/modifier/pyproject.toml
[project]
name = "modifier"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "common",
    "fastapi>=0.121.0",
    "pydantic-settings>=2.11.0",
    "uvicorn>=0.38.0",
]

[tool.uv.sources]
common = { workspace = true }

